requirements.txt
# Core dependencies
streamlit>=1.18.0
python-dotenv>=0.19.0
requests>=2.28.0
pydantic>=1.10.0

# Diagram generation
graphviz>=0.20.0
pydot>=1.4.2
networkx>=2.8.0
matplotlib>=3.6.0
plotly>=5.10.0

# Animation
streamlit-agraph>=0.0.40
streamlit-elements>=0.1.0
streamlit-lottie>=0.0.3
plotly-express>=0.4.0

# UI enhancements
streamlit-option-menu>=0.3.2
streamlit-extras>=0.2.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=5.0.0